requirements.txt
spacy
fastembed
scikit-learn
numpy
feedparser
crawl4ai
xgboost
qdrant-client
google-generativeai
